+++ packages/logic-core/src/data/solver/universal/universalWorker.ts
@@ -167,6 +167,11 @@ onmessage = e => {
   let count = 0;
 
   solve(grid, solution => {
+    if (solution) {
+      if (solution.resetTiles().colorEquals(solution)) {
+        solution = null;
+      }
+    }
     postMessage(solution ? Serializer.stringifyGrid(solution) : null);
 
     count += 1;