+++ packages/logic-core/src/data/solver/z3/z3Solver.ts
@@ -17,8 +17,12 @@ import { array } from '../../dataHelper.js';
 export default class Z3Solver extends Solver {
   public readonly id = 'z3';
 
+  public readonly author = 'Lysine';
+
   public readonly description =
     'Good for confirming that a solution is unique, especially for larger puzzles. It is otherwise slower than most solvers in small to medium-sized puzzles.';
+
+  public readonly supportsCancellation = false;
 
   public async isEnvironmentSupported(): Promise<boolean> {
     try {